WIP fixes some failing tests

+++ db/migrate/20181128031323_create_invoice_items.rb
@@ -2,7 +2,7 @@ class CreateInvoiceItems < ActiveRecord::Migration[5.1]
   def change
     create_table :invoice_items do |t|
       t.integer :quantity
-      t.decimal :unit_price
+      t.integer :unit_price
       t.references :invoice, foreign_key: true
       t.references :item, foreign_key: true
